+++ .swiftlint.yml
@@ -563,7 +563,7 @@ type_body_length:
   error: 400 # error
 file_length:
   warning: 300
-  error: 600
+  error: 1000
 function_body_length:
   warning: 100
   error: 150
